# Conflicts:
#	client/app/scripts/directives/form-builder.js

+++ client/app/greyscale.tables/scripts/services/greyscale-project-products.js
@@ -45,6 +45,7 @@ angular.module('greyscale.tables')
             show: true,
             sortable: 'surveyId',
             dataFormat: 'option',
+            cellTemplate: '<span ng-if="option.id">{{option.title}} <small>(<span ng-show="option.isDraft" translate="SURVEYS.IS_DRAFT"></span><span ng-show="!option.isDraft" translate="SURVEYS.IS_COMPLETE"></span>)</small></span>',
             //dataRequired: true,
             dataSet: {
                 getData: _getSurveys,
@@ -57,17 +58,11 @@ angular.module('greyscale.tables')
                 state: 'projects.setup.surveys.edit({projectId: item.projectId, surveyId: item.surveyId})'
             }
         }, {
-            field: 'surveyStatus',
-            sortable: 'surveyStatus',
-            title: tns + 'SURVEY_STATUS',
-            show: true,
-            dataHide: true
-        }, {
             field: 'workflow.name',
             sortable: 'workflow.name',
             title: tns + 'WORKFLOW',
             show: true,
-            cellTemplate: '{{cell||\'' + tns + 'CREATE_WORKFLOW' + '\'|translate}}',
+            cellTemplate: '{{cell}}<span ng-if="!cell" class="action" translate="' + tns + 'CREATE_WORKFLOW"></span>',
             link: {
                 handler: _editProductWorkflow
             },
@@ -108,14 +103,12 @@ angular.module('greyscale.tables')
             show: true,
             dataFormat: 'action',
             actions: [{
-                title: '',
                 icon: 'fa-pencil',
-                class: 'info',
+                tooltip: 'COMMON.EDIT',
                 handler: _editProduct
             }, {
-                title: '',
                 icon: 'fa-trash',
-                class: 'danger',
+                tooltip: 'COMMON.DELETE',
                 handler: _removeProduct
             }]
         }];
@@ -132,7 +125,6 @@ angular.module('greyscale.tables')
             formWarning: _getFormWarning,
             pageLength: 10,
             add: {
-                title: 'COMMON.CREATE',
                 handler: _editProduct
             }
         };
@@ -192,11 +184,18 @@ angular.module('greyscale.tables')
         }
 
         function _removeProduct(product) {
-            greyscaleProductApi.delete(product.id)
-                .then(_reload)
-                .catch(function (err) {
-                    inform.add('Product delete error: ' + err);
-                });
+            greyscaleModalsSrv.confirm({
+                message: tns + 'DELETE_CONFIRM',
+                product: product,
+                okType: 'danger',
+                okText: 'COMMON.DELETE'
+            }).then(function () {
+                greyscaleProductApi.delete(product.id)
+                    .then(_reload)
+                    .catch(function (err) {
+                        inform.add('Product delete error: ' + err);
+                    });
+            });
         }
 
         function _reload() {
